+++ kfft_math.c
@@ -61,8 +61,8 @@ void
 kfft_math_transpose_scalar(const kfft_scalar* Fin, kfft_scalar* Fout, const uint32_t x,
                            const uint32_t y) {
     for (uint64_t n = 0; n < x * y; n++) {
-        uint32_t i = n / y;
-        uint32_t j = n % y;
+        uint64_t i = n / y;
+        uint64_t j = n % y;
         Fout[n] = Fin[x * j + i];
     }
 }
